feat(1585): PR remarks
